+++ tests/inputs-editorconfig/global/tab-single.lua
@@ -3,5 +3,3 @@ local foo = {
 }
 
 local bar = ""
-
-unpack {}